requirements.txt
pillow>=9.0.0
pdfplumber>=0.7.0
PyMuPDF>=1.20.0
pytesseract>=0.3.10
scikit-learn>=1.0.0
python-dotenv>=0.19.0
